pyproject.toml
[project]
name = "wildlifeml"
version = "0.1.0"
description = "Machine learning project to extract sex (male/female) and age (young/old) characteristics of certain species using camera trap images."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "albumentations>=2.0.5",
    "fastparquet>=2024.11.0",
    "fire>=0.7.0",
    "ipykernel>=6.29.5",
    "numpy<2.0.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "poethepoet>=0.34.0",
    "pre-commit>=4.2.0",
    "pyarrow>=20.0.0",
    "ruff>=0.11.6",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
    "tensorflow==2.10.0",
    "tomli>=2.2.1",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["wildlifeml"]

[tool.ruff]
line-length = 99
target-version = "py310"
lint.select = ["E", "F", "I", "N", "W", "B", "UP", "PL", "RUF"]
lint.ignore = ["PLR2004", "PLR0913"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.poe.tasks.configure-dirs]
cmd = "mkdir -p $WORKING_DIR/configs $WORKING_DIR/logs $WORKING_DIR/data $WORKING_DIR/models"
args = [
    {name="WORKING_DIR", options=["-wd", "--working-dir"], default=".", help="Path to wd containing data and results directories. Uses cwd if no arg passed."},
]

# Run pipeline from top-level wildlife-age-sex directory

[tool.poe.tasks.run-pipeline]
shell = """
    uv run scripts/preprocess.py --config $CONFIG_PATH && \
    uv run scripts/train.py --config $CONFIG_PATH && \
    uv run scripts/evaluate.py --config $CONFIG_PATH
"""
args = [
    {name="CONFIG_PATH", options=["-c", "--config"], default="configs/test__config.toml", help="Path to config file relative to wd."}
]



